# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: Simulacao Credito Creditas
  description: Simulação de empréstimos
  version: 1.0.0
paths:
  /simulacao/credito:
    get:
      summary: Calcular uma simulação de um empréstimo para o cliente
      operationId: getSimulacaoEmprestimo
      tags:
        - Simulacao Credito
      parameters:
        - $ref: '#/components/parameters/cpf'
        - $ref: '#/components/parameters/email'
        - $ref: '#/components/parameters/valorEmprestimo'
        - $ref: '#/components/parameters/dataNascimento'
        - $ref: '#/components/parameters/parcelas'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulacaoEmprestimoResponse'

  /simulacao/cliente:
    get:
      summary: Busca as simulações de empréstimo do cliente
      operationId: getSimulacoesClientes
      tags:
        - Simulacao Credito
      parameters:
        - $ref: '#/components/parameters/cpf'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimulacaoEmprestimoResponse'

  /simulacao/clientes:
    post:
      summary: Realizar processamento de muitas simulações de empréstimos de diversos clientes
      operationId: gerarSimulacoes
      tags:
        - Simulacao Credito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SimulacaoEmprestimoRequest'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimulacaoEmprestimoResponse'


components:
  parameters:

    cpf:
      in: query
      name: cpf
      required: true
      schema:
        type: string

    valorEmprestimo:
      in: query
      name: valorEmprestimo
      required: true
      schema:
        type: number


    parcelas:
      in: query
      name: qtdParcelas
      required: true
      schema:
        type: integer

    dataNascimento:
      in: query
      name: dataNascimento
      required: true
      schema:
        type: string
        format: date
        example: 2000-01-17

    email:
      in: query
      name: email
      required: true
      schema:
        type: string



  schemas:
    SimulacaoEmprestimoResponse:
      type: object
      properties:
        valorEmprestimo:
          type: number
          example: 100.0
        valorTotal:
          type: number
          example: 100.0
        qtdParcelas:
          type: number
          example: 2
        valorParcelas:
          type: number
          example: 10.1
        totalJuros:
          type: number
          example: 101.0
        dataSimulacao:
          type: string
          example: 2025-01-07

    SimulacaoEmprestimoRequest:
      type: object
      properties:
        cpf:
          type: string
          example: "123456"
        email:
          type: string
          example: "a@a.com"
        valorEmprestimo:
          type: number
          example: 100.0
        dataNascimento:
          type: string
          format: date
          example: "2024-10-01"
        qtdParcelas :
          type: integer
          example: 10